Move last parts from mixins to utils and delete mixins
